Create root disk as part of launch request
Previously the disk was created beforehand in a call directly to the
disk resource. But for some reason the disk was not being created
correctly and the kernel refused to boot, claiming no device found.
It should be possible to fix this another way, but this was the
call recommended by the GCE docs.